[ios] convert timestamps to long before converting to date

+++ app-ios/tutanota/Sources/Alarms/TUTAlarmModel.h
@@ -20,7 +20,7 @@ NS_ASSUME_NONNULL_BEGIN
                       repeatPerioud:(TUTRepeatPeriod)repeatPeriod
                            interval:(NSInteger)interval
                             endType:(TUTRepeatEndType)endType
-                           endValue:(NSInteger)endValue
+                           endValue:(long long)endValue
                             trigger:(NSString *)trigger
                       localTimeZone:(NSTimeZone *)localtimeZone
                       scheduleAhead:(NSInteger)scheduleAhead
